+++ src/pages/LoadingScreen.tsx
@@ -3,7 +3,7 @@ import { ActivityIndicator } from "react-native";
 import { Container } from "../components/grid/Container";
 import { MainContainer } from "../components/grid/MainContainer";
 
-export const LoadingScreen = () => {
+export const LoadingScreen: React.FC = () => {
   return (
     <MainContainer>
       <Container isCentered mt={400}>
